add favorites

+++ resources/views/microposts/microposts.blade.php
@@ -12,13 +12,18 @@
                 <div>
                     <p>{!! nl2br(e($micropost->content)) !!}</p>
                 </div>
-                <div>
+                <span>
                     @if (Auth::user()->id == $micropost->user_id)
-                        {!! Form::open(['route' => ['microposts.destroy', $micropost->id], 'method' => 'delete']) !!}
-                            {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-xs']) !!}
-                        {!! Form::close() !!}
+                        @if (!Auth::user()->is_favorites($micropost->id))
+                            {!! Form::open(['route' => ['microposts.destroy', $micropost->id], 'method' => 'delete']) !!}
+                                {!! Form::submit('Delete', ['class' => 'btn btn-danger btn-xs']) !!}
+                            {!! Form::close() !!}
+                        @endif
                     @endif
-            </div>
+                </span>
+                <span>
+                     @include('favorites.favorite_button', ['micropost' => $micropost])
+                </span>
             </div>
         </li>
     @endforeach
